Unified dev and prod webpack configurations
